backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
email-validator==2.1.0
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
openai==1.12.0
anthropic==0.18.1
redis==5.0.1
